# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Unified Holdings Tracker API",
    "version": "1.0.0",
    "description": "统一持仓追踪后端服务 API 文档",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/market/quote": {
      "get": {
        "tags": ["Market"],
        "summary": "获取行情数据",
        "parameters": [
          {
            "name": "codes",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "资产代码列表，逗号分隔",
            "example": "sh600519,hk00700"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回行情数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Quote" }
                }
              }
            }
          },
          "400": {
            "description": "缺少必需参数"
          }
        }
      }
    },
    "/market/kline": {
      "get": {
        "tags": ["Market"],
        "summary": "获取K线数据",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "example": "sh600519"
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["daily", "weekly", "monthly"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "fq",
            "in": "query",
            "schema": { "type": "string", "enum": ["qfq", "hfq", "none"] }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回K线数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/KlinePoint" }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Quote": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "name": { "type": "string" },
          "currentPrice": { "type": "number" },
          "change": { "type": "number" },
          "changePercent": { "type": "number" },
          "volume": { "type": "number" }
        }
      },
      "KlinePoint": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "open": { "type": "number" },
          "close": { "type": "number" },
          "high": { "type": "number" },
          "low": { "type": "number" },
          "volume": { "type": "number" }
        }
      }
    }
  }
}